resolving conflicts and refactoring
