+++ .gitignore
@@ -1,10 +1,4 @@
-.idea/
 .secret*
--# Elastic Beanstalk Files
--*/.elasticbeanstalk/*
--!.elasticbeanstalk/*.cfg.yml
--!.elasticbeanstalk/*.global.yml
-
 *.sw?
 *.egg
 *.egg-info
@@ -22,16 +16,8 @@ doc/build/*
 dist/
 .idea/
 playbooks/*.retry
-
-# Data Files
 *.csv
 *.db
-
-# Elastic Beanstalk Files
-.elasticbeanstalk/*
-!.elasticbeanstalk/*.cfg.yml
-!.elasticbeanstalk/*.global.yml
-
 .vagrant/
 Vagrantfile.local
 aws.yml.local
