+++ docker/requirements.txt
@@ -1,18 +1,20 @@
 slackclient
+python-dateutil
 django-meetup
-click==6.7
-configparser==3.5.0
-dominate==2.3.1
-flake8==3.5.0
-Flask==1.0.2
-Flask-WTF==0.14.2
-itsdangerous==0.24
-Jinja2==2.10
-MarkupSafe==1.0
-mccabe==0.6.1
-pycodestyle==2.3.1
-pyflakes==1.6.0
-visitor==0.1.3
-WTForms==2.2
-flask-mysqldb==0.2.0
-passlib==1.7.1
+click
+configparser
+dominate
+Flask
+Flask-WTF
+itsdangerous
+Jinja2
+MarkupSafe
+mccabe
+pycodestyle
+pyflakes
+visitor
+WTForms
+flask-mysqldb
+passlib
+meetup-api
+slacker
