--- provisioning/playbooks/site.yml
@@ -1,56 +1,0 @@
----
-- name: Deploy Software
-  hosts: all
-  gather_facts: yes
-  become: yes
-
-  roles:
-    - role: geerlingguy.docker
-    - role: geerlingguy.mysql
-    - role: realsalmon.mysql-client
-
-  tasks:
-
-    - name: Create database
-      mysql_db:
-        state: import
-        name: all
-        target: "{{ homedir }}/create_db.sql"
-
-    - name: Update APT package manager repositories cache
-      apt:
-        update_cache: yes
-
-    - name: install "{{ packages }}"
-      package:
-        name: "{{ packages }}"
-        state: present
-      vars:
-        packages:
-          - python3
-          - python3-pip
-
-    - name: install and upgrade via pip
-      pip:
-        name:
-          - pip
-          - virtualenv
-        extra_args: --upgrade
-        executable: pip3
-
-    - name: Install requirments into virtualenv
-      pip:
-        requirements: "{{ homedir }}/requirements.txt"
-        virtualenv: "{{ homedir }}/venv"
-        virtualenv_command: virtualenv
-        virtualenv_python: python3
-
-    - meta: flush_handlers
-
-    - name: start application in background
-      shell: |
-        cd {{ homedir }}
-        . venv/bin/activate
-        python3 application.py {{ mysql_bind_address }} \
-          {{ mysql_root_username }} {{ mysql_root_password }} \
-           </dev/null >/dev/null 2>&1 &
