--- publish/requirements.txt
@@ -1,2 +1,0 @@
-meetup.api
-slackClient
